backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
boto3==1.34.160
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.8.2
email-validator==2.2.0
slowapi==0.1.9
starlette==0.37.2
